Added an option to conditionally execute batches (#33)
* Added an option to conditionally execute batches

* if flags are not provided then all batch elements will
be executed
* Otherwise, each team evaluates its flag and makes
a decision whether or not to perform computations
* Note, it is logical to have a vector of `bool`s for the
flags. However, we can assume that a user wants to compute
how many elements were executed afterwards. Probably,
he/she is going to perform some kinda reduction. Therefore,
they may want to use `unsigned` (even though allowed values
are 0 and 1) because it will be easy for them to evaluate
the sum

* Fixed a bug related to mat-vec operation

* Bumped version 0.0.206